% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_parameter_chunk_size.R
\name{calc_chunk_size}
\alias{calc_chunk_size}
\title{calculate the full chunk size}
\usage{
calc_chunk_size(ds)
}
\arguments{
\item{ds}{dataset}
}
\description{
calculate the full chunk size
}
