% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_parameter_chunk_size.R
\name{calc_parameter_chunk_size}
\alias{calc_parameter_chunk_size}
\title{Calculate the parameter chunk size in bytes}
\usage{
calc_parameter_chunk_size(ds)
}
\arguments{
\item{ds}{data set with one of classes \code{"parameter"}, \code{"text"} or "\code{data}"}
}
\value{
Number of bytes as length 1 numeric vector
}
\description{
Calculate the parameter chunk size in bytes
}
