% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{assign the classes to the dataset list}
\usage{
create_dataset(header_data)
}
\arguments{
\item{header_data}{list of the header information for each block}
}
\description{
assign the classes to the dataset list
}
