% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus.R
\name{get_raw_size.file}
\alias{get_raw_size.file}
\title{method to get the raw size of a file}
\usage{
\method{get_raw_size}{file}(dsn)
}
\arguments{
\item{dsn}{data source name; can be a file path to an OPUS file or directly a
raw vector}
}
\description{
method to get the raw size of a file
}
