% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus_impl.R
\name{get_raw_size.raw}
\alias{get_raw_size.raw}
\title{method to get the raw size of a raw vector}
\usage{
\method{get_raw_size}{raw}(dsn)
}
\arguments{
\item{dsn}{data source name; can be a file path to an OPUS file or directly a
raw vector}
}
\description{
method to get the raw size of a raw vector
}
