% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus_impl.R
\name{get_raw_size}
\alias{get_raw_size}
\title{Dispatch method for get_raw_size}
\usage{
get_raw_size(dsn)
}
\arguments{
\item{dsn}{data source name; can be a file path to an OPUS file or directly a
raw vector}
}
\description{
Dispatch method for get_raw_size
}
