% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus.R
\name{new_list_opusreader2}
\alias{new_list_opusreader2}
\title{Construct new class \code{list_opusreader2}}
\usage{
new_list_opusreader2(dataset_list)
}
\arguments{
\item{dataset_list}{dataset list, where each list element is a measured
spectrum}
}
\description{
Construct new class \code{list_opusreader2}
}
\keyword{internal}
