% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus_impl.R
\name{open_connection.file}
\alias{open_connection.file}
\title{method to open the connection for an opus file}
\usage{
\method{open_connection}{file}(dsn)
}
\arguments{
\item{dsn}{data source name; can be a file path to an OPUS file or directly a
raw vector}
}
\description{
method to open the connection for an opus file
}
