% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus.R
\name{open_connection.raw}
\alias{open_connection.raw}
\title{method to open the connection directly to a raw vector}
\usage{
\method{open_connection}{raw}(dsn)
}
\arguments{
\item{dsn}{data source name; can be a file path to an OPUS file or directly a
raw vector}
}
\description{
method to open the connection directly to a raw vector
}
