% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus.R
\name{open_connection}
\alias{open_connection}
\title{Dispatch method for the open_connection}
\usage{
open_connection(dsn)
}
\arguments{
\item{dsn}{data source name; can be a file path to an OPUS file or directly a
raw vector}
}
\description{
Dispatch method for the open_connection
}
