% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opus_test_files.R
\name{opus_file}
\alias{opus_file}
\title{Get location of a sample OPUS file}
\usage{
opus_file()
}
\value{
a character vector storing the location of the sample OPUS file
}
\description{
Utility function that retrieves the location of the sample OPUS
binary file on disk.
}
\examples{
fn <- opus_file()
fn
}
