% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus.R
\name{opus_lapply}
\alias{opus_lapply}
\title{wrapper function to apply the read_opus_impl() function to a list of
data source paths}
\usage{
opus_lapply(dsn, data_only)
}
\arguments{
\item{dsn}{data source name. Can be a path to a specific file or a path to a
directory. The listing of the files in a directory is recursive.}

\item{data_only}{read data and parameters with \code{FALSE} per default, or only
read data
\code{NULL}, which only returns the parsed data as an in-memory R object.}
}
\description{
wrapper function to apply the read_opus_impl() function to a list of
data source paths
}
