% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opus_test_dsn.R
\name{opus_test_dsn}
\alias{opus_test_dsn}
\title{Get File Paths of Sample OPUS Files Included in the Package}
\usage{
opus_test_dsn()
}
\value{
a character vector with the paths OPUS files included in the package
}
\description{
Utility function that retrieves the location of the sample OPUS
binary file on disk.
}
\examples{
(dsn <- opus_test_dsn)
}
