% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opus_test_dsn.R
\name{opus_test_file}
\alias{opus_test_file}
\title{Get path of a Selected Sample OPUS File Included in the Package}
\usage{
opus_test_file()
}
\value{
a character vector with the path to a selected single sample OPUS
file
}
\description{
Utility function that retrieves the location of the sample OPUS
binary file on disk.
}
\examples{
(fn <- opus_test_file())
}
