% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_chunk.R
\name{parse_chunk.data}
\alias{parse_chunk.data}
\title{read chunk method for data}
\usage{
\method{parse_chunk}{data}(ds, raw)
}
\arguments{
\item{ds}{chunk dataset}

\item{raw}{raw vector}
}
\description{
read chunk method for data
}
\keyword{internal}
