% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_chunk.R
\name{parse_chunk.parameter}
\alias{parse_chunk.parameter}
\title{read chunk method for parameter}
\usage{
\method{parse_chunk}{parameter}(ds, con)
}
\arguments{
\item{ds}{chunk dataset}

\item{con}{connection to raw vector}
}
\description{
read chunk method for parameter
}
