% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_chunk.R
\name{parse_chunk}
\alias{parse_chunk}
\title{read chunk method}
\usage{
parse_chunk(ds, raw)
}
\arguments{
\item{ds}{chunk dataset}

\item{raw}{raw vector}
}
\description{
read chunk method
}
\seealso{
Parsing helper
\code{\link{parse_header}()}
}
\concept{parsing}
\keyword{internal}
