% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_chunk.R
\name{parse_chunk.text}
\alias{parse_chunk.text}
\title{read chunk method for text}
\usage{
\method{parse_chunk}{text}(ds, raw)
}
\arguments{
\item{ds}{chunk dataset}

\item{raw}{raw vector}
}
\description{
read chunk method for text
}
\keyword{internal}
