% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.list_opusreader2}
\alias{print.list_opusreader2}
\title{Print method for collection of OPUS spectra with class \code{list_opusreader2}}
\usage{
\method{print}{list_opusreader2}(x, ...)
}
\arguments{
\item{x}{List with OPUS spectra collection of class \code{list_opusreader2}}

\item{...}{Additional arguments passed to print method}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print method for collection of OPUS spectra with class \code{list_opusreader2}
}
