% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bin_types.R
\name{read_character}
\alias{read_character}
\title{read character from binary}
\usage{
read_character(con, cursor, n = 1L)
}
\arguments{
\item{con}{connection to raw vector}

\item{cursor}{offset}

\item{n}{number of elements}
}
\description{
read character from binary
}
\seealso{
Other read_bytes: 
\code{\link{read_double}()},
\code{\link{read_float}()},
\code{\link{read_signed_int}()},
\code{\link{read_unsigned_int}()}
}
\concept{read_bytes}
