% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bin_types.R
\name{read_double}
\alias{read_double}
\title{read double from binary (double-precision, 64 bits)}
\usage{
read_double(con, cursor, n = 1L)
}
\arguments{
\item{con}{connection to raw vector}

\item{cursor}{offset}

\item{n}{number of elements}
}
\description{
read double from binary (double-precision, 64 bits)
}
\seealso{
Other read_bytes: 
\code{\link{read_character}()},
\code{\link{read_float}()},
\code{\link{read_signed_int}()},
\code{\link{read_unsigned_int}()}
}
\concept{read_bytes}
