% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bin_types.R
\name{read_float}
\alias{read_float}
\title{read float from binary (single-precision, 32 bits)}
\usage{
read_float(con, cursor, n = 1L)
}
\arguments{
\item{con}{connection to raw vector}

\item{cursor}{offset}

\item{n}{number of elements}
}
\description{
read float from binary (single-precision, 32 bits)
}
\seealso{
Other read_bytes: 
\code{\link{read_character}()},
\code{\link{read_double}()},
\code{\link{read_signed_int}()},
\code{\link{read_unsigned_int}()}
}
\concept{read_bytes}
