% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus_file.R
\name{read_opus_file}
\alias{read_opus_file}
\title{Read OPUS binary file of a Bruker spectrometer}
\usage{
read_opus_file(file)
}
\arguments{
\item{file}{character vector with the path to the OPUS file}
}
\value{
list containing the different data chunks of an OPUS file, possibly:
- \code{refl_data_param}:
- \code{refl}:
- \code{quant_report_refl}:
- \code{sc_sample_data_param}:
- \code{sc_sample}:
- \code{sc_ref_data_param}:
- \code{sc_ref}:
- \code{optics}:
- \code{optics_ref}:
- \code{acquisition_ref}:
- \code{fourier_transformation_ref}:
- \code{fourier_transformation}:
- \code{sample}:
- \code{acquisition}:
- \code{ìnstrument_ref}:
- \code{instrument}:
- \code{lab_and_process_param_1}:
- \code{lab_and_process_param_2}:
- \code{ìnfo_block}:
- \code{history}:
}
\description{
This function can be used to read and parse an OPUS file,
to make it usable for other processing steps.
}
\examples{
library(opusreader2)

file_path <- system.file("extdata/test_data/BF_lo_01_soil_cal.1", package = "opusreader2")

opus_list <- read_opus_file(file_path)

}
