% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus.R
\name{read_opus_impl}
\alias{read_opus_impl}
\title{Read a single opus file}
\usage{
read_opus_impl(dsn, data_only)
}
\arguments{
\item{dsn}{source path of an opus file}

\item{data_only}{read data and parameters with \code{FALSE} per default, or only
read data}
}
\description{
Read a single opus file
}
