% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus.R
\name{read_opus_parallel_mirai}
\alias{read_opus_parallel_mirai}
\title{Read chunks of OPUS files in parallel using \code{{mirai}} backend via
\code{mirai::mirai_map()}}
\usage{
read_opus_parallel_mirai(dsn, data_only, progress_bar)
}
\arguments{
\item{dsn}{data source name. Can be a path to a specific file or a path to a
directory. The listing of the files in a directory is recursive.}

\item{data_only}{read data and parameters with \code{FALSE} per default, or only
read data \code{NULL}, which only returns the parsed data.}

\item{progress_bar}{print a progress bar. Default is \code{FALSE}.}
}
\description{
Relies on background processes (daemons) set up via \code{mirai::daemon()}
}
\keyword{internal}
