% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus.R
\name{read_opus_single}
\alias{read_opus_single}
\title{Read a single OPUS file}
\usage{
read_opus_single(dsn, data_only = FALSE)
}
\arguments{
\item{dsn}{source path of an opus file}

\item{data_only}{read data and parameters with \code{FALSE} per default, or only
read data}
}
\value{
list with parsed OPUS measurement and data blocks of the spectrum
measurement contained in the OPUS file. See \code{\link[=read_opus]{read_opus()}} for details on the
list elements returned (first-level block names and information provided).
}
\description{
Read a single OPUS file
}
