% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus_raw.R
\name{read_raw.file}
\alias{read_raw.file}
\title{method to open the connection for an opus file}
\usage{
\method{read_raw}{file}(dsn)
}
\arguments{
\item{dsn}{source path of an opus file}
}
\description{
method to open the connection for an opus file
}
