% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus_raw.R
\name{read_raw}
\alias{read_raw}
\title{Dispatch method for the open_connection}
\usage{
read_raw(dsn)
}
\arguments{
\item{dsn}{source path of an opus file}
}
\description{
Dispatch method for the open_connection
}
\seealso{
Connection helper
\code{\link{set_connection_class}()}
}
\concept{connection}
