% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bin_types.R
\name{read_unsigned_int}
\alias{read_unsigned_int}
\title{read unsigned integer from binary}
\usage{
read_unsigned_int(con, cursor, n = 1L)
}
\arguments{
\item{con}{connection to raw vector}

\item{cursor}{offset}

\item{n}{number of elements}
}
\description{
read unsigned integer from binary
}
\seealso{
Other read_bytes: 
\code{\link{read_signed_int}()}
}
\concept{read_bytes}
