% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seek_opus.R
\name{seek_opus}
\alias{seek_opus}
\title{reposition cursor for open read connection}
\usage{
seek_opus(con, cursor)
}
\arguments{
\item{con}{connection to raw vector}

\item{cursor}{cursor position within the connection}
}
\description{
reposition cursor for open read connection
}
