% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus_impl.R
\name{set_connection_class}
\alias{set_connection_class}
\title{define class of dsn}
\usage{
set_connection_class(dsn)
}
\arguments{
\item{dsn}{data source name; can be a file path to an OPUS file or directly a
raw vector}
}
\description{
define class of dsn
}
