% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_opus.R
\name{open_connection}
\alias{open_connection}
\title{Dispatch method for the open_connection}
\usage{
open_connection(dsn)
}
\arguments{
\item{dsn}{data source name; can be a file path to an OPUS file or directly a
raw vector}
}
\description{
Dispatch method for the open_connection
}
\seealso{
Connection helpers
\code{\link{get_raw_size}()},
\code{\link{set_connection_class}()}
}
\concept{connection}
