% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus.R
\name{opus_lapply}
\alias{opus_lapply}
\title{List wrapper that reads a list of files (\code{dsn}) via \code{read_opus_single()}
and returns spectra in a list}
\usage{
opus_lapply(dsn, data_only)
}
\arguments{
\item{dsn}{data source name. Can be a path to a specific file or a path to a
directory. The listing of the files in a directory is recursive.}

\item{data_only}{read data and parameters with \code{FALSE} per default, or only
read data
\code{NULL}, which only returns the parsed data as an in-memory R object.}
}
\value{
nested list where first-level elements are individual spectral
measurements parsed from individual OPUS binary files; the output is
identical to its user exposed reading interface, see \code{?read_opus}

spectra list containing the elements described in \code{?read_opus}
}
\description{
List wrapper that reads a list of files (\code{dsn}) via \code{read_opus_single()}
and returns spectra in a list
}
\seealso{
\code{\link[=read_opus]{read_opus()}} \code{\link[=read_opus_single]{read_opus_single()}}
}
