% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_chunk.R
\name{parse_chunk.default}
\alias{parse_chunk.default}
\title{read chunk method for default}
\usage{
\method{parse_chunk}{default}(ds, con)
}
\arguments{
\item{ds}{chunk dataset}

\item{con}{connection to raw vector}
}
\description{
read chunk method for default
}
\keyword{internal}
