% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_chunk.R
\name{parse_chunk.parameter}
\alias{parse_chunk.parameter}
\title{read chunk method for parameter}
\usage{
\method{parse_chunk}{parameter}(ds, raw)
}
\arguments{
\item{ds}{chunk dataset}

\item{raw}{raw vector}
}
\description{
read chunk method for parameter
}
\keyword{internal}
