% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_header.R
\name{parse_header}
\alias{parse_header}
\title{parse the header of the opus file}
\usage{
parse_header(raw)
}
\arguments{
\item{raw}{raw vector of the opus binary file}
}
\description{
parse the header of the opus file
}
\seealso{
Parsing helper
\code{\link{parse_chunk}()}
}
\concept{parsing}
\keyword{internal}
