% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_header.R
\name{parse_header}
\alias{parse_header}
\title{parse the header of the opus file}
\usage{
parse_header(raw_size, con)
}
\arguments{
\item{raw_size}{raw vector of the opus binary file}

\item{con}{connection to the raw vector}
}
\description{
parse the header of the opus file
}
