% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_opus.R
\name{parse_opus}
\alias{parse_opus}
\title{Parse data and parameters of OPUS binary file of a Bruker spectrometer}
\usage{
parse_opus(dsn, data_only)
}
\arguments{
\item{dsn}{data source name; can be a file path to an OPUS file or directly a
raw vector}

\item{data_only}{TRUE or FALSE if only the ab or refl data should be returned}
}
\value{
list containing the different data chunks of an OPUS file, possibly:
\itemize{
\item \code{refl_data_param}:
\item \code{refl}:
\item \code{quant_report_refl}:
\item \code{sc_sample_data_param}:
\item \code{sc_sample}:
\item \code{sc_ref_data_param}:
\item \code{sc_ref}:
\item \code{optics}:
\item \code{optics_ref}:
\item \code{acquisition_ref}:
\item \code{fourier_transformation_ref}:
\item \code{fourier_transformation}:
\item \code{sample}:
\item \code{acquisition}:
\item \code{instrument_ref}:
\item \code{instrument}:
\item \code{lab_and_process_param_1}:
\item \code{lab_and_process_param_2}:
\item \code{info_block}:
\item \code{history}:
}
}
\description{
This function can be used to read and parse OPUS files,
to make it usable for other processing steps.
}
\examples{
library(opusreader2)

dsn <- system.file("extdata/test_data/BF_lo_01_soil_cal.1", package = "opusreader2")

opus_list <- parse_opus(dsn, data_only = FALSE)
}
\seealso{
Core functions
\code{\link{read_opus}()}
}
\concept{core}
