% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bin_types.R
\name{read_character}
\alias{read_character}
\title{read character from binary}
\usage{
read_character(con, cursor, n = 1L, encoding = "latin1")
}
\arguments{
\item{con}{connection to raw vector}

\item{cursor}{offset}

\item{n}{number of elements}
}
\description{
read character from binary
}
