% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bin_types.R
\name{read_double}
\alias{read_double}
\title{read double from binary (double-precision, 64 bits)}
\usage{
read_double(raw, cursor, n = 1L)
}
\arguments{
\item{raw}{raw vector}

\item{cursor}{offset}

\item{n}{number of elements}
}
\description{
read double from binary (double-precision, 64 bits)
}
