% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bin_types.R
\name{read_float}
\alias{read_float}
\title{read float from binary (single-precision, 32 bits)}
\usage{
read_float(con, cursor, n = 1L)
}
\description{
read float from binary (single-precision, 32 bits)
}
