% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bin_types.R
\name{read_float}
\alias{read_float}
\title{read float from binary (single-precision, 32 bits)}
\usage{
read_float(raw, cursor, n = 1L)
}
\arguments{
\item{raw}{raw vector}

\item{cursor}{offset}

\item{n}{number of elements}
}
\description{
read float from binary (single-precision, 32 bits)
}
\keyword{internal}
