% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus.R
\name{read_opus}
\alias{read_opus}
\alias{read_opus_single}
\title{Read OPUS binary file of a Bruker spectrometer}
\usage{
read_opus(dsn, progress = !(length(dsn) == 1))

read_opus_single(dsn)
}
\arguments{
\item{dsn}{data source name; can be a file path to an OPUS file or directly a
raw vector}

\item{progress}{Logical, whether a progress bar message is printed. Defaults to \code{TRUE} if several data sources are passed to the \code{dsn} parameter, \code{FALSE} otherwise.}
}
\value{
list containing the different data chunks of an OPUS file, possibly:
\itemize{
\item \code{refl_data_param}:
\item \code{refl}:
\item \code{quant_report_refl}:
\item \code{sc_sample_data_param}:
\item \code{sc_sample}:
\item \code{sc_ref_data_param}:
\item \code{sc_ref}:
\item \code{optics}:
\item \code{optics_ref}:
\item \code{acquisition_ref}:
\item \code{fourier_transformation_ref}:
\item \code{fourier_transformation}:
\item \code{sample}:
\item \code{acquisition}:
\item \code{instrument_ref}:
\item \code{instrument}:
\item \code{lab_and_process_param_1}:
\item \code{lab_and_process_param_2}:
\item \code{info_block}:
\item \code{history}:
}
}
\description{
This function can be used to read and parse an OPUS file,
to make it usable for other processing steps.
}
\examples{
library(opusreader2)

dsn <- system.file("extdata/test_data/BF_lo_01_soil_cal.1", package = "opusreader2")

opus_list <- read_opus(dsn)

# Multiple DSN can be passed to `read_opus`

opus_list <- read_opus(rep(dsn, times = 10))

}
