% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus.R
\name{read_opus}
\alias{read_opus}
\title{Read OPUS binary file of a Bruker spectrometer}
\usage{
read_opus(dsn, data_only = FALSE, parallel = FALSE, progress_bar = FALSE)
}
\arguments{
\item{dsn}{data source name. Can be a path to a specific file or a path to a
directory. The listing of the files in a directory is recursive.}

\item{data_only}{read data and parameters with \code{FALSE} per default, or only
read data
\code{NULL}, which only returns the parsed data as an in-memory R object.}

\item{parallel}{read files in parallel via chunking. Default is \code{FALSE}.}

\item{progress_bar}{print a progress bar. Default is \code{FALSE}.}
}
\description{
This function can be used to read and parse OPUS files,
to make it usable for other processing steps.
}
\seealso{
Core functions
\code{\link{parse_opus}()}
}
\concept{core}
