% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus_file.R
\name{read_opus_file}
\alias{read_opus_file}
\title{Read OPUS binary file of a Bruker spectrometer}
\usage{
read_opus_file(file, progress = !(length(file) == 1))
}
\arguments{
\item{file}{character vector with the path to the OPUS file(s)}

\item{progress}{Logical (defaults to \code{TRUE} if several files are passed to the \code{file} parameter, \code{FALSE} otherwise), whether a progress bar message is printed}
}
\value{
list containing the different data chunks of an OPUS file, possibly:
\itemize{
\item \code{refl_data_param}:
\item \code{refl}:
\item \code{quant_report_refl}:
\item \code{sc_sample_data_param}:
\item \code{sc_sample}:
\item \code{sc_ref_data_param}:
\item \code{sc_ref}:
\item \code{optics}:
\item \code{optics_ref}:
\item \code{acquisition_ref}:
\item \code{fourier_transformation_ref}:
\item \code{fourier_transformation}:
\item \code{sample}:
\item \code{acquisition}:
\item \code{instrument_ref}:
\item \code{instrument}:
\item \code{lab_and_process_param_1}:
\item \code{lab_and_process_param_2}:
\item \code{info_block}:
\item \code{history}:
}
}
\description{
This function can be used to read and parse an OPUS file,
to make it usable for other processing steps.
}
\examples{
library(opusreader2)

file <- system.file("extdata/test_data/BF_lo_01_soil_cal.1", package = "opusreader2")

opus_list <- read_opus_file(file)
}
