% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus_file.R
\name{read_opus_file}
\alias{read_opus_file}
\title{Read OPUS binary file of a Bruker spectrometer}
\usage{
read_opus_file(file)
}
\arguments{
\item{file}{character vector with the path to the OPUS file}
}
\value{
list containing the different chunks of an OPUS file
}
\description{
This function can be used to read and parse an OPUS file,
to make it usable for other processing steps.
}
\examples{
library(opusreader2)

file_path <- system.file("extdata/test_data/BF_lo_01_soil_cal.1", package = "opusreader2")

opus_list <- read_opus_file(file_path)

}
