% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus_raw.R
\name{read_opus_raw}
\alias{read_opus_raw}
\title{Read the raw vector of an opus data source}
\usage{
read_opus_raw(dsn)
}
\arguments{
\item{dsn}{source path of an opus file}
}
\value{
list of raw vectors
}
\description{
Read the raw vector of an opus data source
}
