% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus.R
\name{read_opus_single}
\alias{read_opus_single}
\title{Read a single OPUS file}
\usage{
read_opus_single(dsn, data_only = FALSE)
}
\arguments{
\item{dsn}{source path of an opus file}

\item{data_only}{read data and parameters with \code{FALSE} per default, or only
read data}
}
\description{
Read a single OPUS file
}
