% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bin_types.R
\name{read_signed_int}
\alias{read_signed_int}
\title{read signed integer from binary}
\usage{
read_signed_int(raw, cursor, n = 1L)
}
\arguments{
\item{raw}{raw vector}

\item{cursor}{offset}

\item{n}{number of elements}
}
\description{
read signed integer from binary
}
\keyword{internal}
