% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bin_types.R
\name{read_unsigned_int}
\alias{read_unsigned_int}
\title{read unsigned integer from binary}
\usage{
read_unsigned_int(raw, cursor, n = 1L)
}
\arguments{
\item{cursor}{offset}

\item{n}{number of elements}

\item{con}{connection to raw vector}
}
\description{
read unsigned integer from binary
}
