% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_opus_raw.R
\name{set_connection_class}
\alias{set_connection_class}
\title{define class of dsn}
\usage{
set_connection_class(dsn)
}
\arguments{
\item{dsn}{source path of an opus file}
}
\description{
define class of dsn
}
\keyword{internal}
