% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_opus.R
\name{set_connection_class}
\alias{set_connection_class}
\title{define class of dsn}
\usage{
set_connection_class(dsn)
}
\arguments{
\item{dsn}{data source name; can be a file path to an OPUS file or directly a
raw vector}
}
\description{
define class of dsn
}
\seealso{
Connection helpers
\code{\link{get_raw_size}()},
\code{\link{open_connection}()}
}
\concept{connection}
